Windows:
- Replicated logic to actually enable Total Shield block-all firewall rules only when VPN is connected/connecting.

- Added again the logic to disable IPv6 when Total Shield gets enabled.
